# shared/api/http_openapi.yaml
openapi: 3.0.0
info:
  title: Intrusion Detection API
  version: 1.0.0
  description: Backend and Detection Service API contracts

paths:
  /api/models:
    get:
      summary: List all models
      responses:
        '200':
          description: List of models
    post:
      summary: Upload new model
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [coco, face, fire, custom]
      responses:
        '200':
          description: Model uploaded
    put:
      summary: Update model settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelUpdate'
      responses:
        '200':
          description: Model updated

  /api/zones:
    get:
      summary: List all zones
      responses:
        '200':
          description: List of zones
    post:
      summary: Create zone
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Zone'
      responses:
        '200':
          description: Zone created
    put:
      summary: Update zone
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Zone'
      responses:
        '200':
          description: Zone updated
    delete:
      summary: Delete zone
      parameters:
        - name: zone_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Zone deleted

  /api/events:
    get:
      summary: List events with filters
      parameters:
        - name: zone
          in: query
          schema:
            type: string
        - name: cls
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
        - name: t_start
          in: query
          schema:
            type: number
        - name: t_end
          in: query
          schema:
            type: number
      responses:
        '200':
          description: List of events

  /api/query:
    post:
      summary: Chatbot query
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                t_start:
                  type: number
                t_end:
                  type: number
      responses:
        '200':
          description: Query results

  /api/sos:
    post:
      summary: Trigger or cancel SOS
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [trigger, cancel]
      responses:
        '200':
          description: SOS state updated

  /detector/start:
    post:
      summary: Start detection stream
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                source:
                  type: object
                  properties:
                    type:
                      type: string
                      enum: [rtsp, usb, file]
                    uri:
                      type: string
                models:
                  type: array
                  items:
                    $ref: '#/components/schemas/ModelConfig'
                zones_version:
                  type: string
      responses:
        '200':
          description: Detection started

  /detector/stop:
    post:
      summary: Stop detection
      responses:
        '200':
          description: Detection stopped

  /detector/status:
    get:
      summary: Get detection status
      responses:
        '200':
          description: Status information

components:
  schemas:
    ModelUpdate:
      type: object
      properties:
        enabled:
          type: boolean
        conf:
          type: number
        iou:
          type: number
        enabled_classes:
          type: object
    ModelConfig:
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
        conf:
          type: number
        iou:
          type: number
        enabled_classes:
          type: object